性能优化
1. State::toString() 字符串拼接优化
    使用 std::ostringstream 替代多次字符串拼接
2. 循环中重复创建随机数分布对象
    添加 _uniformFloatDist 成员变量
3. 重复的 weak_ptr.lock() 调用优化
    在 threadModelStorage 中只调用一次 lock()
    提高线程安全性
4. 动作选择算法优化
    使用累积分布函数（CDF）和二分查找
    复杂度从 O(n) 降至 O(log n)
5. 修复循环中使用值传递而非引用
    将值传递改为引用传递，避免不必要的拷贝
6. 移除未使用的变量
7. 优化文件 I/O 异常处理
    使用 is_open() 检查文件状态
    添加读取完整性检查
8. 优化模型保存使用临时文件
    先写入临时文件，成功后再原子替换
    避免保存失败时损坏原文件
9. 优化 getOperateOpt 方法
10. 优化 dynamic_pointer_cast 使用

+++ native/desc/Widget.cpp
@@ -34,9 +34,9 @@ namespace fastbotx {
         if (STATE_WITH_TEXT || Preference::inst()->isForceUseTextModel()) {
             bool overMaxLen = this->_text.size() > STATE_TEXT_MAX_LEN;
             this->_text = this->_text.substr(0, STATE_TEXT_MAX_LEN * 4);
-            int cutLength = STATE_TEXT_MAX_LEN;
+            size_t cutLength = static_cast<size_t>(STATE_TEXT_MAX_LEN);
             if (this->_text.length() > cutLength && isZhCn(this->_text[STATE_TEXT_MAX_LEN])) {
-                int ci = 0;
+                size_t ci = 0;
                 for (; ci < cutLength; ci++) {
                     if (isZhCn(this->_text[ci])) {
                         ci += 2;
